Check Media Duration Before Download
Check media duration before downloading a file using `pyffmpeg`.

I will look into using `pyffmpeg` for our other ffmpeg functions but it will require a bit of work. The library does not bundle ffprobe, and instead just uses ffmpeg's log file to achieve the features that ffprobe would normally give you.

It seems to be made more so for simple usage by just running `FFmpeg().convert()`. Regardless, it does provide us with a bundled ffmpeg so it is worth switching the version that we use.